Add `nvim` shortcuts for tabs
This commit updates NeoVim's config file in following ways:
 - Enables `Ctrl+t` to open new tab and `Ctrl+w` to close it
 - Enables `Ctrl+Left/Right` to switch to previous/next tab, and
   analogously for `Crtl+j/k`

Note the this nicely integrates with the vim support in VS Code. Because
current vim config uses `<C-t>` to toggle NerdTree, the keybinding for
'New Text File ...' stopped working in VS Code. But with this change it
picks up the `<C-t>` for the same purpose.

Additionally, `Ctrl+Right` does not break the visual user interaction of
`Crtl+Tab` in VS Code which is also nice.