Fix install options of the `fzf` vim plugin
This commit changes the command that's executed to install/update the
`fzf` plugin in Neovim.

Previously the install script generated `~/.fzf.{bash,zsh}`
configuration files in addition to downloading the binary. This was
unsatisfactory for two reasons:
 1. The configuration was already done by the fzf omz plugin
 2. Placing these files in `$HOME` does not conform to XDG specification

An alternative to this change would be to add the `--xdg` option as well
to generate these files into `$XDG_CONFIG_HOME/fzf`.