Install Gnome Alacritty toggle extension
This commit replaces the X11-only `alacritty_toggle.sh` script by Gnome
extension `alacritty-toggle` which works both on X11 and Wayland.

The extension is cloned from a forked repository for the additional
support of Gnome 42 and custom Alacritty installation path.

The installation can be done via `make alacritty-toggle`, however, the
extension is not picked up by the current Gnome session. So one must
 1. logout and log back in
 2. re-run the make target

Hence the extra make target instead of extending `make alacritty`.

Additionally, `alacritty_toggle.sh` was deprecated as well as any other
code that was mentioning it (such as dconf files and make targets), and
was removed.