Fix Alacritty on Wayland with Nvidia
This commit adds .zprofile with a hotfix for running Alacritty terminal
on Wayland with Nvidia.

Specifically, this change sets the `__EGL_VENDOR_LIBRARY_FILENAMES`
environment variable. It is essential to use mesa for the EGL vendor
library, otherwise Alacritty crashes on startup.

Note that for this change to take an effect, one must restart current
session.

Additionally, this commit also adds a README section referencing how to
enable Wayland on a machine with Nvidia.