Disable `~/.sudo_as_admin_successful`
This commit adds a make target `disable-sudo-admin-file` which generates
custom sudoers config file that disables the `admin_flag`.

The effect of this change is that `.sudo_as_admin_successful` should
no longer appear in HOME. This solves the respective part of issue #17.

Note that this solution only works with `sudo >= 1.9.6` and that there
are no additional steps necessary, just run the make target and remove
the file from HOME.