Fix config path for `nvidia-settings` desktop
This commit makes the `nvidia-settings` application use custom location
of the configuration file when running it from the Gnome menu.

Specifically this patch:
 - Adds and exports `NVIDIA_SETTINGS_RC` environment variable
 - Adds new make target `nvidia-settings-rc-xdg-path` which edits the
   `Exec` entry in `/usr/share/applications/nvidia-settings.desktop` to
   use the same `--config` option as the zsh alias
 - Adds new section describing this manual configuration to README

In order to apply these changes just reload current shell and run the
newly added make target.

Note that this fixes the remaining issues with `nvidia-settings` and
concludes this part of issue #17.