Switch to official repository of alacritty toggle
Before this commit the `alacritty-toggle` make target cloned a fork of
the `gnome-alacritty-toggle` repository which added support for Gnome
42.

Because relevant PR, adding this support to the official repository, has
been merged, there's no need to keep using the fork. Therefore, this
commit updates the repository URL in Makefile and relevant docs.