Fix ripgrep colors in nvim search
This commit changes the `--color` option in ripgrep's config file to
`auto` and adds some documentation from the official FAQ page.

Previously, `rg` used `--colors=always`, resulting in additional
"coloring" symbols in its output. This did not work particularly well in
Neovim, where the search command `<leader> s` would not expect colored
output from ripgrep. The result would be broken, hardly readable, and
consequently unusable search result.

The result of this change is that now `vim-ripgrep` works as expected,
and when `rg` is used in terminal, its output is colored as well.