Rename `base16-project` to `tinted-theming`
This commit renames the `base16-project` git repository and URLs in
documentation to `tinted-theming`.

The project (organisation) has been recently renamed and the *Tinted
Theming* name should be used from now on. See the GitHub repository for
more info on this matter.

Note that all the old URLs have been redirected and continue to work,
but it's definitely safer to make the switch now.

There's no need for migration, except maybe to reinstall the vim plugin
(just to be consistent):
 1. fetch these changes
 2. temporarily comment out the plugin
 3. run `:PlugClean` to remove the old plugin
 4. re-enable new plugin and re-install with `:PlugUpdate`