Uninstall previous binenv distribution version
This commit extends the binenv part of the `upgrade` script by adding a
cleanup procedure.

The commit adds a `binenv_clean` function which is called in the
distribution loop after an update (install). This cleanup function then
searches for the last non-default version (i.e. the supposedly
previously installed version) and removes it.

Known issues:
 - If certain distribution has no recent update (i.e. no new version has
   been installed), the `binenv uninstall` will display an annoying
   error to stdout. The command will succeed but the log message would
   better be skipped.
 - It would be more sound to uninsall _all_ the currently installed
   versions which are obsolete, not just to attempt to uninstall the
   latest one (which might not even be installed anyway).