Change `nvim` installation from PPA to Snap
This commit changes the installation method for Neovim form a PPA to
Snap. The reason is that the PPA is not maintained by the official team
behind Neovim and currently contains quite an old stable release
version.

The officially maintained Snap package is not ideal (for instance it
does not ship with man page), it is up to date with the latest release
which actually fixes some recent bugs (for instance W18 warning on
startup when reading the base16 plugin files).

Migration notes:
 1. Uninstall Neovim: `sudo apt purge neovim python3-pynvim`
 2. Cleanup: `sudo apt autoremove`
 3. Remove PPA:
    `sudo add-apt-repository --remove ppa:neovim-ppa/stable`
 4. Re-install Neovim: `make neovim`