Add installation of ClickHouse Client
This commit adds installation of the ClickHouse CLI client as a new
`make clickhouse` target. The installation uses an official download
script and by default installs the binary to `$XDG_BIN_HOME`.

Note that it also sets up the `CLICKHOUSE_HISTORY_FILE` environment
variable to point to a XDG-complient location.