Deprecate `cargo-latest-dirname` script
This commit removes the `cargo-latest-dirname` bash script which was
used as part of the `make rust-tools` recipe to find build artifacts
(specifically completion scripts and man pages) of the latest `cargo
install`.

Instead the tools either:
 - Override the path to which its build script outputs its artifacts
   (e.g., `bat`)
 - Or uses the source directory that `cargo` downloads into its registry

In either case, the `cargo-latest-dirname` is no longer necessary. Also,
these new methods should be more reliable in picking the correct
version and should not fail because the build has not produced anything.