Replace auxiliary `~/.zshenv`
This commit removes (or rather replaces) the auxiliary `.zshenv` script
which was symlinked directly under `$HOME`.

This was unsatisfactory w.r.t. the XDG directory spec, even more so that
the sole purpose of the script was to export `ZDOTDIR` and source the
actual config file.

So instead of doing that there is new systemd user-specific config under
`.config/environment.d/` which is loaded during session login and thus
_before_ shell login, and which now configures the `ZDOTDIR`.

Note that these conf files are loaded and merged in the lexicographic
order of their names (hence the choice of `00-default.conf`). Also, any
change to these files will typically need a new login to be caught and
applied (user unit can be updated using
`systemctl daemon-reload --user`, but that change won't affect services
started _before_ - e.g., terminal, VMs, etc.).

An alternative solution would be to use `.pam_environment`, but a) it's
directly under HOME too, and b) it's been deprecated.

Finally, in order to migrate just
 1. Run `make links`
 2. Remove the deprecated `~/.zshenv`
 3. Re-login