Install `linux-tools` and set perf `buildid.dir`
This commit adds new `make kernel-tools` recipe (dependency of `make
basic-tools`), which installs kernel version-specific tools such as
`perf`, and sets its `buildid.dir`.

This change is in an attempt to get rid of `~/.debug`. Note that the
config is currently set just for `--system`, so it's a question, whether
if invoked under current USER it's gonna work. The issue with using
`--user` option is that it creates `~/.perfconfig`, which defeats the
purpose here.