Migrate Neovim options and key maps to Lua
This commit moves all Neovim configs that are not directly related to
plugins to Lua.

Originally pretty much all the nvim configs were located in `init.vim`.
With this path, most configs were translated to Lua and organized in
several modules (inspired by the structure of the `nvchad` project):
 - `options.lua` contains both global and non-global _options_ (that is,
   configs set via `vim.opt`, `vim.o`, and `vim.g`)
 - `mappings.lua` contains all the key bindings set via `vim.keymap.set`
 - `autocmds.lua` is dedicated to autocommand setup

Note that `init.vim` is still necessary for the plugin setup (and also
for some plugin-related config). It is, however, planned to migrate
plugins to `lazy.nvim` and replace the init file with `init.lua`.

Migration notes:
 - Since there are new (lua) files, these should be symlinked via
   `make links`
 - The directory for the `spell`-related files has been changed to the
   default location (`~/.config/nvim/spell/`), so one should move it
   there from the old location (`~/.local/state/vimdid`)
 - The `scripts/spacetab.vim` script has been removed, so the symlink
   (and probably the whole scripts dir) can also be dropped

Known issues & limitations:
 - Plugins are still managed from `init.vim` (vim-plug)
 - Options and autocommands related to a specific plugin (or LSP) should
   be lazily applied during plugin setup
 - Moving multi-line selection with `A-{j,k}` is bugged (moving _single_
   line around works tho)
 - `lsp_signature` reports an error on startup in a Rust project with a
   lot of dependencies (but works afterwards)