Deprecate `vim-plug` in favor of `lazy.nvim`
This commit finishes the transition of Neovim configs to Lua by removing
remaining `vim-plug` plugins (LSP-based completions) and adding them to
`lazy.nvim` setup.

Moreover, this change finally enables getting rid of `init.vim`. So
starting with this patch, _all_ of Neovim's config is based on Lua
implementations.

Note that the migration plan here is:
 1. Symlink new sources
 2. Purge old vim-plug directory and dead symlinks

Finally, note that although plugins are now managed by `lazy.nvim`, this
does _not_ mean that all the plugins are Lua-based. The transition so
far was 1:1 and there might be future patches that switch some plugins
to their Lua-based counterparts.