Switch from `fzf` to `skim`
This commit add a manual installation of `skim` (a Rust-based fuzzy
finder) and uses it in most places where `fzf` was used (zsh, nvim).

For now this is a try out setup and `fzf` is still installed.
Furthermore, note that the shell colors set by `base16-fzf` are reused
for skim as well via `FZF_DEFAULT_OPTS`.

There are also some minor tweaks compared to the previous setup:
 - The search window height has been slightly increased to 25%
 - Searching in shell now displays a small preview window (using `bat`)

Known issues & limitations:
 - When searching in Neovim, the colors are currently not being
   correctly propagated by either the Base16 plugin or fzf-lua