Add custom `zsh-virsh` plugin
This commit add a custom zsh plugin for `virsh` extensions. Currently
there is a single helper function `virsh-rm-pool`.

Future commits might clean up `.zshrc` in a similar fashion. That is, it
would be nice to move application-specific configs from the rc file to
dedicated plugins, so they are contained and can be easily ignored
(e.g., when certain binary is not present).