Install Alacritty desktop assets without sudo
This commit changes the location of Alacritty's desktop assets to user's
directories (under XDG_DATA_HOME). This removes the requirement to
install these with sudo (i.e., globally).

Also, `/usr/share/pixmaps/` is slowly being deprecated in favor of the
"icons" dirs. Finally, note that the terminfo as well as the apt
dependencies still require sudo.