Replace `pyright` with `pylsp`
Pyright does not quite do the job for me, plus it used to be installed
globally via npm. Therefore, this commit replaces it with a
community-managed python-lsp-server (`pylsp`).

Note that `ruff` is now installed as part of `python-lsp-server`, that
is it's no longer in its own virtual environment, but rather injected
into `python-lsp-server` (aside from that it's still available as a
binary).

In order to migrate:
 1. Install pylsp via `make python-lsp-server`
 2. Uninstall pyright via `npm uninstall -g pyright`