Lower the frequency of snap refreshes
This commit adds a new `make snap` recipe to adjust snap settings.
Specifically, it lowers the frequency of updates to once a week and
disables them while an app is running.

The rationale behind this change is that I do prefer to run update
manually via the `upgrade` script and regular snap update notifications
became annoying. Another option would be to actually hold all the snaps,
but this should be a good compromise.