Fix `zsh-sdkman` plugin updates
This commit fixes `make sdk` when it's used to update an existing
installation.

Previously this target would fail because of an already cloned directory
of the `zsh-sdkman` plugin. Now, it will pull latest changes in that
directory instead of trying to clone it again.

Note that the installing/updating the plugin has been extracted from the
`sdk` target to a new recipe `zsh-sdkman` (called as part of `zsh`).