Remove Zed editor
This commit removes `make zed` since I've not used it outside of just
experimenting with the editor. It's a good editor, but I do prefer the
simple nature of Neovim and living in the terminal.

Note that in order to uninstall Zed, just run `zed --uninstall` and
follow the instructions.