Migrate from `omz` to `zinit`
This commit contains a major change to the zsh setup, plugin, and
completion management. It switches from oh-my-zsh to zinit.

The main motivation is that zinit is more lean and thus faster. When
benchmarked (via `time  zsh -i -c exit`) it's approximately twice faster
than the previous (omz) setup. Moreover, there's still room for
improvement as this is is just a "MVP".

Another fact that justifies this transition is that oh-my-zsh plugins
and library functions can still be used via `OMZP` and `OMZL`
plugins/snippets from `zinit`.

Note that some tools that are only relevant in the interactive shell
(e.g. `forgit`) or plugins (e.g., `zsh-sdkman`) were moved from Makefile
to zshrc.

Finally, note that as part of this change the `upgrade omz` has been
deprecated in favor of `upgrade zsh`.