Format and lint shell scripts and rc files
This commit includes a re-formatting of all the shell scripts and
bash/zsh rc files. Additionally, it applies small stylistic code changes
that fix `shellcheck` errors/warnings.

Furthermore, new `.editorconfig` entries were added for Makefile,
Python, and various shells (i.e., file types present in this repo).

Note that the `.pre-commit-config.yaml` was
 1. Updated, so that each hook uses the latest version
 2. Extended with a new shellcheck hook to lint all the shell files

Also note that currently `shfmt` will fail to parse `.zshrc` due to one
zsh-specific line in `zshaddhistory`. This prevents the whole file to be
formatted - this should be fixed in future commits.

Finally, the `shfmt` argument list in neovim's formatter was updated to
by default conform to Google's style guide (that is, to indent with two
spaces).