Move `zshaddhistory` to `history.zsh`
This commit sets up new directory for zsh functions `$ZSH_FUNCTIONS` and
moves `zshaddhistory` definition to `$ZSH_FUNCTIONS/history.zsh`.

The motivation for this change is that `shfmt` does not support zsh
shell and, unfortunately, `zshaddhistory` is zsh-specific. The problem
here is that `shfmt` fails to parse the whole `.zshrc` file and thus
does not work there just because of one function. Hence, we move it to a
dedicated `history.zsh` script and source it from zshrc.