Add `starship` as an alternative prompt
This commit adds an alternative Starship prompt setup to `.zshrc`.

The `starship` installation is done by downloading its binary from
GitHub releases and creating a symlink to it.

Creating a symlink for zinit managed binaries is done via the
`zinit-annex-binary-symlink` extension. Also note that the `$ZPFX/bin`
(directory containing linked or compiled artifacts) is added to PATH.

Finally, note that there is no customization for the starship prompt
yet, so it will use its default look & feel.