Generate static pandoc and pipx completions
This commit removes the pandoc and pipx zinit plugins which used to
generate and source respective zsh completions at init. Instead, the
completion files are generated when installed/updated in Makefile and
stored in `$ZSH_COMPLETIONS`.

The problem with the way these zinit plugins worked was two-fold:
 1. They used to fail because they required compinit to be run _before_,
    but it's delayed to a single call _after_. Note that this was due to
    the fact that these were `as'null'` rather than `as'completion'`.
 2. The completion script had to be generated on each new shell (or more
    precisely on each init), however, these completions can only change
    when pandoc/pipx is updated (which is less frequent).

Similarly, the aws-vault completion snippet has been fixed - it is now
configured `as'completion'`, which means that `compdef` will be deferred
until `cdreplay`.