Pre-generate `stack` bash completions
This commit removes the `OMZP::stack` plugins and replaces it with
custom (null) one, which pre-generates its bash completions and loads
them from a static file.

Note that the `OMZP::stack` plugins was quite slow (comparatively),
which was most likely because it runs `bashcompinit`. This is not
necessary, since at the point the plugin was loaded these was already
done.