Setup zsh module for auto compiling sourced files
This commit builds and loads new zsh module for automatically loading
sourced files as part of zshrc.

Note that currently the `zdharma-continuum/zinit-module` needs to be
patched for zsh version 5.8.1 and later, which is done by newly added
script `zinit-module-build`.

The zinit-module-build script:
 1. Downloads a git patch from a zinit-module forked repository which
    contains the compatibility fix
 2. Validates its integrity
 3. Applies the patch to the cloned zinit-module repository
 4. Builds the zinit-module - essentially what the `zinit module build`
    (resp., `mod-install.sh`) does

Note that the awkward patch fetching and validation is because the patch
contains some trailing whitespace and if it had been stored in this
repository, it would not be reproducible since the pre-commit would
"fix" it and break it. For a long term solution it would be better to
make own patch/fork, but hopefully this gets into upstream soon.