Install nodejs via nvm
This commit reworks the way nodejs is installed in the Makefile from an
installation script from GitHub to a more official way via `nvm`.

The nvm manager is currently installed also via an installation script
from the official GH repo (it'd be nice to change this to `git clone`)
and is installed to `$NVM_DIR` (`$XDG_DATA_HOME/nvm/`).

`zshrc` now also has `OMZP::nvm` added to source the nvm initialization
script (in a lazy way; note: node/npm binaries are no longer
automatically added to the PATH unless being in an interactive shell).

Nodejs is now installed via nvm under a specific version in the NVM_DIR.
Furthermore, there is new `make typescript` recipe to install the
TypeScript compiler and LSP.