Rollback to previous non-breaking nvim-cmp commit
This is a hotfix for a backwards-incompatible change in nvim-cmp (and
related plugins). This commit rolls back to and pins the last commit
working in neovim v0.10.x.

Note that this is a temporary fix, a proper upstream fix hopefully comes
soon. Also note that this issue should also be resolved in neovim
v0.11.0, but that upgrade brings other issues (e.g., with inlayhints).