Update LSP config for Neovim v0.11.0
This commit adds two changes to Neovim's LSP config:
 1. Reverts commit pinning for nvim-cmp
 2. Switches inlay hint plugin from `lsp-inlayhints` which is archived
    and not working in v0.11.0 to `nvim-lsp-endhints`

Note that the new inlay hint plugin is not a one-to-one replacement.
Specifically, it does not support displaying the hits only for the
current line. However, it at least displays the hints at the end of each
line instead of the complete visual clutter from the built-in LSP hint
support.