Lighten starship prompt
This commit reworks the starship prompt configuration.

This new config is inspired by the pure prompt except that here we push
environment & language modules to the right of the first prompt line.
Also, currently the git state module is disabled (compared to pure).

Additionally, the gruvbox color palette has been removed and the styling
relies on the standard terminal color names, which makes it work for
other tinty themes than just gruvbox-dark-hard.